% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C-series.r
\name{c2}
\alias{c2}
\title{put a value into position 2 of an existing vector of length 1}
\usage{
c2(x, position2x)
}
\arguments{
\item{x}{the exist vector of length 1}

\item{position2x}{the value you want to put into position 2 of the vector}
}
\description{
put a value into position 2 of an existing vector of length 1
}
